version fonctionnelle mais pas opti (9,77 s)

+++ shouka/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: armendes <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/24 09:46:07 by armendes          #+#    #+#             */
-/*   Updated: 2020/08/25 11:10:24 by armendes         ###   ########.fr       */
+/*   Updated: 2020/08/25 15:10:49 by armendes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,16 +21,16 @@
 
 typedef struct	s_list
 {
-	int				x;
-	int				y;
+	long				x;
+	long				y;
 	struct s_list	*next;
 	struct s_list	*next_line;
 }				t_list;
 
 typedef struct	s_param
 {
-	int				xmax;
-	int				ymax;
+	long				xmax;
+	long				ymax;
 	char			plein;
 	char			vide;
 	char			obstacle;
@@ -38,21 +38,21 @@ typedef struct	s_param
 	struct s_list	*next;
 }				t_param;
 
-t_list		*create_elem(int x, int y);
+t_list		*create_elem(long x, long y);
 void		add_elem(t_list *elem1, t_list *elem2);
 void		add_next_line(t_list *elem1, t_list *elem2);
 void		put_next_line(t_list *begin_list);
 t_list		*list_last(t_list *begin_list);
-int			stock(char c, t_param *params, int *count, int i, int j);
-int			check_nb_col(int fd, int max, t_param *params);
-t_param		*get_map_info(int fd, char *name);
+long			stock(char c, t_param *params, long *count, long i, long j);
+long			check_nb_col(long fd, long max, t_param *params);
+t_param		*get_map_info(long fd, char *name);
 t_param		*split_info(char *str, char *name);
-void		disp(int xpos, int ypos, t_param *elem, int i);
-int			check_obst(int x, int y, t_list *begin_list, int i);
-void		solve(int xbord, int ybord, t_param *begin_list, int i);
+void		disp(long xpos, long ypos, t_param *elem, long i);
+long			check_obst(long x, long y, t_list *begin_list, long i);
+void		solve(long xbord, long ybord, t_param *begin_list, long i);
 void		ft_putchar(char c);
-int			ft_strlen(char *str);
-int			ft_atoi(char *str);
+long			ft_strlen(char *str);
+long			ft_atoi(char *str);
 
 
 #endif
