+++ shouka/srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: armendes <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/24 11:53:49 by armendes          #+#    #+#             */
-/*   Updated: 2020/08/25 13:09:42 by armendes         ###   ########.fr       */
+/*   Updated: 2020/08/25 15:11:35 by armendes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 #include <stdio.h>
 
-int		main(int ac, char **av)
+int		main(int	ac, char **av)
 {
-	int			i;
-	int			fd;
+	long			i;
+	long			fd;
 	t_param		*params;
 
 	i = 1;
